Factor out non-CourseScheduler state dependent methods to utils
